Moved duplicated whitehall image stubbing code to FixturesHelper
To save creating a new helper file just for one method, this
seems a good place to keep the code. It is helping stub
network requests to return fixture data - so in that respect
is a fixture helper, even though it's somewhat different to
the other methods in this file.

In the next commit we'll be testing the import of multiple
images, which will require stubbing multiple times, which is
why it makes sense to encapsulate this stubbing in one place.