Change UniqueFilenameService to take hash parameters
We will soon be adding parameters to the UniqueFilenameService
to make it more configurable. Some parameters can be omitted
and the order shouldn't matter, making this ideal to be converted
to hash parameters.

I've also reordered the parameters so that 'original_filename'
is the first parameter, as it is the most important. (We won't
necessarily have to pass 'ensure_unique_against' in future, but
we'll always need a base filename to work against).