WIP: Abort if one of the later images is not 960x640
We want to 'abort early' if one of the images later on is
not of the right size, to save having to roll back our changes later.

As a result, we now dowload all images in an edition ahead of time, so
that we can check their dimensions. This means that some of the fixtures
used earlier, which were never evaluated before, now need to be stubbed.

In stubbing the gif, I discovered that animated gifs don't work with the
ImageNormaliser - nor are they expected to ('Please upload a gif that is
not animated') - and returns 'nil'. We should probably double check that
the gif we are importing is not animated, since otherwise the import
script throws a runtime error rather than a handled exception.

And in the spirit of aborting early before encountering a late image that
is not valid, we should apply this principle across editions. Currently
we only check images 'in advance', so to speak, on a per-edition basis,
meaning we can get into a semi-imported state before we discover a bad
image and have to abort.

Finally, this commit tests the exception message explicitly - we should
be doing this for every abort exception in my opinion, as we now have
multiple ways in which the exception can be raised.