Add test helper for Edition assertions
At the moment we use edition assertions as a means to abort a request
whenever a user is interacting with something that is not in the valid
state for the scenario the user is in.

We previously haven't done any testing of these, which is a bit of a
shame since they're quite important. However it's difficult to do these
tests concisely. To try counter this I've put together a shared helper
that can be used without too much set-up and applied to request tests.