Add request tests for Contact Embed
As part of this a few small issues were discovered:

When API is down we render a 200 response, this has been changed to a
503 response. Likely this was done so that the modal could show the
issue but that status code was misrepresentative of the situation. This
change will mean that the modal falls back to the default modal error,
but I think that is fine.

We were returning a text/html response when just sending text back. This
is changed to text/plain.

We were unnecessarily setting a layout attribute when it can only be
application.