Add request tests for Publish
This resolves a few issues in the PublishController and it removes the
dead end feature tests that were wrote for publish.

The only casualty of this seems to be the test where we verify that the
flash has been created where it might be better to assert the text
appears in the element. We could achieve this by either using
assert_select, pulling in Capybara helpers or using https://github.com/kucaahbe/rspec-html-matchers
which is how I approached this at first.