Add request tests for Documents
This adds requests test for the DocumentsController. It fixes a couple
of minor issues where we weren't asserting edition state and were using
an implicit redirect.

The feature tests that are a dead end have been removed. As part of this
we've lost an assertion that previous value is shown to the user. This
is something that we could consider adding at request or view level (it
probably sits more naturally at view level) however it isn't something
tested on most other forms so felt like an anomaly and should be
considered as something to add as part of a broader strategy.