WIP - replace mutating document_import with a context object
Our problem is that we mutate the document_import object during the
import, this conflicts with our all or nothing strategy for
transactions. What happens is that we create a document and various
other models in the database and then, when a transaction fails, roll
them back. However as we set the document attribute on a document_import
the rolled back model still exists in memory and when we save the
document_import it tries to save the document model meaning we at best
get an exception and at worst get a half created document.

The remedy to this is to not mutate document_import until we are
satisfied that the document has been fully imported, and so do this at
the last step so it can be closer to being atomic.

We've experimented in this commit with adding a context model so
document_import isn't passed around.

It might be that we can actually simplify things from this solution
though. We could, for instance, run reload before we do any mutating of
a whitehall_import so that we've cleaned out on any values stored in
memory. Example: https://github.com/alphagov/content-publisher/compare/alternative-atomic-imports?expand=1

Couple of other tweaks this work did:

- Move setting a document as imported to be in the transaction, that way
if this step raises an exception we haven't persisted a document but
failed to set the status.
---

Working out how to test it?
- are we over mocking?

Why do we pass around the document_import and have to set document on it first?

Import
- creating document

CreateEdition
- needs to know the document to create the edition

CreateRevision
- Only needs to know the document import to pass for attachments

CreateImageRevision (same in CreateFileAttachmentRevision)
- We iterate through every asset Whitehall has told it has for that edition
- We then decide whether we want to remove those assets (no longer needed) or redirect them to Content Publishers new assets
- We store these decisions in WhitehallMigration::AssetImport models

Maybe we also want user ids to be a property of our context?

Context itself might be a bad name, what do we call a class that
represents a collection of the migration data?