Prototype of change history/dates to Publishing API payload
This is an example of changes needed so that Content Publisher can pass
through the necessary data to populate the change history and publishing
dates of content.

It is dependent on a number of changes to Content Publisher data
modelling and the expectation that the Publishing API can receive both a
change_history and change_note.

It is designed to support both re-syncing content and initial drafting /
publishing.

As a side-note we may want to move this class outside the namespace of
PreviewDraftEditionService as this is used outside that context in
resyncing.