Add tempfile method for migrated image
CreateFileAttachmentBlobService is used through two mechanisms:
1. uploading a file directly to Content Publisher through the UI, where
file is a ActionDispatch::Http::UploadedFile, which has a tempfile
method
2. through migration from Whitehall, where file is a
WhitehallImporter::AttachmentFileDecorator, which does not have a
tempfile method

Calling the method tempfile in the latter cases results in a no method
error.

Therefore adding a tempfile method which returns the tmpfile for the
attachment.