Get details of the selected option in interactor
It takes that page you are currently on:
`params[:document_type_selection_id]`

and uses that to get the all of the options that can be selected on that
page:
`DocumentTypeSelection.find(params[:document_type_selection_id]).options`

and then uses the selected option:
`params[:selected_option_id]`

to get the hash of the selected option.

e.g. it should return

```
{
  id: "foo",
  type: "managed_elsewhere",
  path: "/foo",
}
```