Consolidate requirement checkers into 'checkers/' subfolder
I found it odd that we had checker_issues.rb and issue.rb in the same
folder as all of the checkers themselves. The checkers embody the
requirements themselves, whereas these two files act as an interface
to those requirements.

Although it involves more code changes to move the checkers into a
subfolder than it does to move the two named files, I think this is
now a more sensible folder structure than before:
- 'requirements/' contains all requirement-related code
- 'requirements/checkers' contains the checkers themselves

I've also moved the corresponding specfiles into a subfolder. This
looks a little odd as we have no tests for checker_issues.rb/issue.rb
directly.