Fix change_history bug
This fixes a bug introduced where an error is thrown if there is no
`change_history` for an edition in the Publishing API and `#length` is called
on it from within ChangeHistoryCheck.

This commit sets the default change_history to an empty array if it's not
present so that calling `#length` on it now returns 0 instead of a method
error.