Handle editions with no change history in Integrity Checker
We have found instances of Whitehall documents with no changes notes; somehow
these were first published in 2016 with no change note. This lack of history is
consistent with the edition representation in the Publishing API, which also
has no change notes. While at least there is consistency in the lack of change
history between Whitehall and Content Publisher, the Integrity Checker fails
for these documents because in the proposed payload we have generated a "First
published." change note which of course doesn't match with what's in the
Publishing API. We've decided it's ok to send the "First published." change
note to Publishing API after these documents have been imported since they
should have had it in the first place and at least it gives the document some
history. As such, we've had to introduce some logic to check that in the event
that the Publishing API has no change history, the proposed payload that has
been generated contains only the "First published." note.