Make change history check more realistic
Previously this code assumed that publishing-api would return
change_history for draft editions. This has turned out to be incorrect
as publishing-api only returns change_history for live editions. This
has lead to some discrepancies when checking the state of change history
between cp and pub-api.

The difference in behaviour is that content-publisher can return
change_history for non-live editions if the draft edition is a major
change with a change_note and it's not the first edition. See ->
https://github.com/alphagov/content-publisher/commit/72b6556e218d66d49e4fd21f14743ea49926d36f#diff-4e6bc683d583eff1852c25cce42073d2R26

The above means that content_history can be different when comparing
drafts so we need to trim the latest draft change_note in cp to reflect
the state in publishing-api.