Use render with no arguments in all view specs
This changes the view specs to do local assignment via RSpec mocking
(which is consistent with the existing approach in spec_helper.rb) which
allows to no longer specify which view needs to be rendered in a spec.
This removes the need for the self.class.top_level_description variable
which was unusual and allows the test descriptions to match the view
path again.

It also adapts the timeline_entry factory to accept an argument for the
contents field which resolves the need to modify this after building the
model.