Use RSpec matchers for attachment meta component
This switches all the test expectations to be done using RSpec style
rather than a mixture of RSpec and Minitest. I imagine we've ended up in
this approach as govuk_publishing_components has lots of these mixtures
however this difference makes this test different to other Content
Publisher tests.

This also removes the `type: :view` declaration which is unnecessary as
this is inferred from the file location.