Remove test that raises warning
Checking whether code raises any error raises a warning with RSpec and
instead a particular exception should be checked. However the code that
is under test is not responsible for raising this exception and to test
which exception would involve coupling ourselves to
govuk_publishing_components internals.

An alternative approach to removing this test could be to raise our own
exception but this doesn't feel necessary.