Convert Requirements::Checker class into a module
This change is similar a change to the previous coomit, where a longer
backstory is provided.

In this commit I'm changing Requirements::Checker from a class to a
module so that it's methods can be mixed into classes as a trait rather
than through sub-class inheritance.

Requirements::Checker is used as a means to share the boilerplate of
an interface to the various checker classes. These checker classes
don't act as sub-types of a Requirements::Checker class as they all
have distinctly different interfaces (and can't be used interchangeably).
Which leads me to conclude that module is the more idiomatic approach
to sharing the code.